requirements.txt
azure-identity
azure-mgmt-media
azure-mgmt-storage
azure-storage-blob
python-dotenv